MR-65: Implemented shared modals/Added confirm delete action

+++ src/Components/SuperAdmins/index.jsx
@@ -1,14 +1,19 @@
+/* eslint-disable no-unused-vars */
 import styles from './super-admins.module.css';
 import Form from './Form/index';
-import MessageModal from './MessageModal/index';
+import ResponseModal from '../Shared/ResponseModal';
 import Table from './Table/index';
 import { useEffect, useState } from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import ConfirmModal from '../Shared/ConfirmModal';
 
 function SuperAdmins() {
   const [superadmins, setSuperadmins] = useState([]);
   const [showModal, setshowModal] = useState(false);
+  const [showConfirm, setShowConfirm] = useState(true);
+  const [idToDelete, setIdToDelete] = useState(null);
   const [resMessage, setResMessage] = useState('');
+  const [state, setSate] = useState('');
 
   const getSuperadmins = async () => {
     try {
@@ -16,8 +21,9 @@ function SuperAdmins() {
       const data = await res.json();
       setSuperadmins(data.data);
     } catch (error) {
+      setSate('fail');
       setResMessage(error.message);
-      setshowModal(true);
+      openModal();
     }
   };
   useEffect(() => {
@@ -30,7 +36,12 @@ function SuperAdmins() {
   const openModal = () => {
     setshowModal(true);
   };
-
+  const closeConfirm = () => {
+    setShowConfirm(false);
+  };
+  const openConfirm = () => {
+    setShowConfirm(true);
+  };
   const addItem = async (superadmin) => {
     try {
       const res = await fetch(`${process.env.REACT_APP_API_URL}/api/super-admins`, {
@@ -42,13 +53,16 @@ function SuperAdmins() {
       });
       getSuperadmins();
       if (res.ok) {
+        setSate('success');
         setResMessage('New superadmin created');
         openModal();
       } else {
+        setSate('fail');
         setResMessage('Failed to create superadmin');
         openModal();
       }
     } catch (error) {
+      setSate('fail');
       setResMessage(error.message);
       openModal();
     }
@@ -64,32 +78,42 @@ function SuperAdmins() {
         body: JSON.stringify(updatedSuperadmin)
       });
       if (res.ok) {
+        setSate('success');
         setResMessage('Superadmin updated');
         openModal();
       } else {
+        setSate('fail');
         setResMessage('Failed to update superadmin');
         openModal();
       }
     } catch (error) {
+      setSate('fail');
       setResMessage(error.message);
       openModal();
     }
   };
-
-  const deleteItem = async (id) => {
+  const confirmDelete = (id) => {
+    openConfirm();
+    setIdToDelete(id);
+  };
+  const deleteItem = async () => {
     try {
-      const res = await fetch(`${process.env.REACT_APP_API_URL}/api/super-admins/${id}`, {
+      const res = await fetch(`${process.env.REACT_APP_API_URL}/api/super-admins/${idToDelete}`, {
         method: 'DELETE'
       });
+      closeConfirm();
       if (res.ok) {
+        setSate('success');
         setResMessage('Superadmin deleted');
         openModal();
-        setSuperadmins([...superadmins.filter((superadmin) => superadmin._id !== id)]);
+        setSuperadmins([...superadmins.filter((superadmin) => superadmin._id !== idToDelete)]);
       } else {
+        setSate('fail');
         setResMessage('Failed to delete superadmin');
         openModal();
       }
     } catch (error) {
+      setSate('fail');
       setResMessage(error.message);
       openModal();
     }
@@ -97,10 +121,21 @@ function SuperAdmins() {
 
   return (
     <section className={styles.container}>
-      {showModal && <MessageModal msg={resMessage} onClose={closeModal} />}
+      {showModal && <ResponseModal state={state} message={resMessage} handler={closeModal} />}
+
+      {showConfirm && (
+        <ConfirmModal
+          title={'Delete superadmin'}
+          reason={'delete'}
+          handler={closeConfirm}
+          onAction={deleteItem}
+        >
+          Are you sure you want to delete this superadmin?
+        </ConfirmModal>
+      )}
       <h2 className={styles.h2}>Superadmin List</h2>
       <Router>
-        <Table data={superadmins} deleteItem={deleteItem} />
+        <Table data={superadmins} deleteItem={deleteItem} confirmDelete={confirmDelete} />
         <Switch>
           <Route path="/super-admins/form/:id">
             <Form putItem={putItem} addItem={addItem} />
